chore(tool): consolidate ast package tests into per-source-file pairing
Second package in the test-consolidation pass for #1128 (tool/ex went
first in #1142). tool/internal/ast had five orphaned/segmented test
files with no direct source-file pairing:

  ast_test.go, funcmatch_test.go, parser_extra_test.go,
  primitives_builders_test.go, shared_extra_test.go,
  shared_helpers_test.go

All content tested parser.go, primitives.go, or shared.go, so each
gets folded into that source file's _test.go:

  {ast_test.go, parser_extra_test.go} -> parser_test.go (new)
  primitives_builders_test.go -> primitives_test.go
  {funcmatch_test.go, shared_extra_test.go, shared_helpers_test.go}
    -> shared_test.go

directive_test.go and typename_test.go already paired correctly and
are untouched. This is a pure move: no test was added, removed,
renamed, or edited, and the {name, pass/fail} set is identical before
and after the merge. Self-referential filename literals in the former
ast_test.go ("ast_test.go" used as a fixture path to parse this
package's own source) were updated to "parser_test.go" to match the
new filename.
