+++ tool/internal/ast/shared_test.go
@@ -395,3 +395,409 @@ type GenMulti[K comparable, V any] struct{ m map[K]V }
 		assert.NotNil(t, FindStructType(src, "GenMulti"))
 	})
 }
+
+// makeFuncDecl builds a minimal *dst.FuncDecl for testing.
+func makeFuncDecl(params, results []*dst.Field) *dst.FuncDecl {
+	ft := &dst.FuncType{}
+	if len(params) > 0 {
+		ft.Params = &dst.FieldList{List: params}
+	}
+	if len(results) > 0 {
+		ft.Results = &dst.FieldList{List: results}
+	}
+	return &dst.FuncDecl{
+		Name: dst.NewIdent("TestFunc"),
+		Type: ft,
+		Body: &dst.BlockStmt{},
+	}
+}
+
+func field(typeExpr dst.Expr) *dst.Field {
+	return &dst.Field{Type: typeExpr}
+}
+
+func ident(name string) *dst.Ident { return &dst.Ident{Name: name} }
+
+func selector(pkg, name string) *dst.SelectorExpr {
+	return &dst.SelectorExpr{X: ident(pkg), Sel: ident(name)}
+}
+
+// mustMatch matches decl against r with no import context (nil), i.e. as if
+// decl's enclosing file were unavailable. Type filters resolve via the
+// import-path-tail fallback in this mode; see TestFuncDeclMatchesFilters_ImportAliasResolution
+// for matching with a real *dst.File's import declarations.
+func mustMatch(t *testing.T, decl *dst.FuncDecl, r *rule.InstFuncRule) bool {
+	t.Helper()
+	ok, err := funcDeclMatchesFilters(decl, r, nil)
+	require.NoError(t, err)
+	return ok
+}
+
+func TestFuncDeclMatchesFilters_NoFilters(t *testing.T) {
+	decl := makeFuncDecl(
+		[]*dst.Field{field(ident("string"))},
+		[]*dst.Field{field(ident("error"))},
+	)
+	r := &rule.InstFuncRule{}
+	assert.True(t, mustMatch(t, decl, r), "no filters should always match")
+}
+
+func TestFuncDeclMatchesFilters_ExactSignature(t *testing.T) {
+	// func(string, int) (float32, error)
+	decl := makeFuncDecl(
+		[]*dst.Field{field(ident("string")), field(ident("int"))},
+		[]*dst.Field{field(ident("float32")), field(ident("error"))},
+	)
+
+	tests := []struct {
+		name string
+		sig  rule.FuncSignature
+		want bool
+	}{
+		{
+			name: "exact match",
+			sig:  rule.FuncSignature{Args: []string{"string", "int"}, Returns: []string{"float32", "error"}},
+			want: true,
+		},
+		{
+			name: "wrong arg type",
+			sig:  rule.FuncSignature{Args: []string{"int", "string"}, Returns: []string{"float32", "error"}},
+			want: false,
+		},
+		{
+			name: "wrong arg count",
+			sig:  rule.FuncSignature{Args: []string{"string"}, Returns: []string{"float32", "error"}},
+			want: false,
+		},
+		{
+			name: "wrong return type",
+			sig:  rule.FuncSignature{Args: []string{"string", "int"}, Returns: []string{"error"}},
+			want: false,
+		},
+		{
+			name: "no args filter only checks returns",
+			sig:  rule.FuncSignature{Returns: []string{"float32", "error"}},
+			want: false, // sig.Args==nil means 0 expected, but decl has 2 params
+		},
+		{
+			name: "args only, no returns check",
+			sig:  rule.FuncSignature{Args: []string{"string", "int"}},
+			want: false, // sig.Returns==nil means 0 expected, but decl has 2 results
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			sig := tt.sig
+			r := &rule.InstFuncRule{Signature: &sig}
+			assert.Equal(t, tt.want, mustMatch(t, decl, r))
+		})
+	}
+}
+
+func TestFuncDeclMatchesFilters_SignatureContains(t *testing.T) {
+	// func(context.Context, string) error
+	decl := makeFuncDecl(
+		[]*dst.Field{field(selector("context", "Context")), field(ident("string"))},
+		[]*dst.Field{field(ident("error"))},
+	)
+
+	tests := []struct {
+		name string
+		sig  rule.FuncSignature
+		want bool
+	}{
+		{
+			name: "arg match triggers true",
+			sig:  rule.FuncSignature{Args: []string{"context.Context"}},
+			want: true,
+		},
+		{
+			name: "return match triggers true",
+			sig:  rule.FuncSignature{Returns: []string{"error"}},
+			want: true,
+		},
+		{
+			name: "no match",
+			sig:  rule.FuncSignature{Args: []string{"int"}, Returns: []string{"bool"}},
+			want: false,
+		},
+		{
+			name: "second arg matches",
+			sig:  rule.FuncSignature{Args: []string{"string"}},
+			want: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			sig := tt.sig
+			r := &rule.InstFuncRule{SignatureContains: &sig}
+			assert.Equal(t, tt.want, mustMatch(t, decl, r))
+		})
+	}
+}
+
+func TestFuncDeclMatchesFilters_Result(t *testing.T) {
+	// func() (io.Reader, error)
+	decl := makeFuncDecl(
+		nil,
+		[]*dst.Field{field(selector("io", "Reader")), field(ident("error"))},
+	)
+
+	assert.True(t, mustMatch(t, decl, &rule.InstFuncRule{Result: "error"}))
+	assert.True(t, mustMatch(t, decl, &rule.InstFuncRule{Result: "io.Reader"}))
+	assert.False(t, mustMatch(t, decl, &rule.InstFuncRule{Result: "io.Writer"}))
+	assert.False(t, mustMatch(t, decl, &rule.InstFuncRule{Result: "string"}))
+}
+
+func TestFuncDeclMatchesFilters_LastResult(t *testing.T) {
+	// func() (io.Reader, error)
+	decl := makeFuncDecl(
+		nil,
+		[]*dst.Field{field(selector("io", "Reader")), field(ident("error"))},
+	)
+
+	// error is the final result
+	assert.True(t, mustMatch(t, decl, &rule.InstFuncRule{LastResult: "error"}))
+	// io.Reader is NOT the final result
+	assert.False(t, mustMatch(t, decl, &rule.InstFuncRule{LastResult: "io.Reader"}))
+}
+
+func TestFuncDeclMatchesFilters_Param(t *testing.T) {
+	// func(context.Context, string) error
+	decl := makeFuncDecl(
+		[]*dst.Field{field(selector("context", "Context")), field(ident("string"))},
+		[]*dst.Field{field(ident("error"))},
+	)
+
+	assert.True(t, mustMatch(t, decl, &rule.InstFuncRule{Param: "context.Context"}))
+	assert.True(t, mustMatch(t, decl, &rule.InstFuncRule{Param: "string"}))
+	assert.False(t, mustMatch(t, decl, &rule.InstFuncRule{Param: "int"}))
+}
+
+func TestFuncDeclMatchesFilters_CombinedFilters(t *testing.T) {
+	// func(context.Context, string) (io.Reader, error)
+	decl := makeFuncDecl(
+		[]*dst.Field{field(selector("context", "Context")), field(ident("string"))},
+		[]*dst.Field{field(selector("io", "Reader")), field(ident("error"))},
+	)
+
+	// All filters match → true
+	sig := rule.FuncSignature{Args: []string{"context.Context", "string"}, Returns: []string{"io.Reader", "error"}}
+	r := &rule.InstFuncRule{
+		Signature:  &sig,
+		Result:     "error",
+		LastResult: "error",
+		Param:      "context.Context",
+	}
+	assert.True(t, mustMatch(t, decl, r))
+
+	// Signature matches but Param doesn't → false
+	r2 := &rule.InstFuncRule{
+		Signature: &sig,
+		Param:     "int",
+	}
+	assert.False(t, mustMatch(t, decl, r2))
+}
+
+func TestFuncDeclMatchesFilters_NoParams(t *testing.T) {
+	// func() error
+	decl := makeFuncDecl(nil, []*dst.Field{field(ident("error"))})
+
+	// Empty signature matches
+	r := &rule.InstFuncRule{Signature: &rule.FuncSignature{Returns: []string{"error"}}}
+	assert.True(t, mustMatch(t, decl, r))
+
+	// Non-empty args don't match
+	r2 := &rule.InstFuncRule{Signature: &rule.FuncSignature{Args: []string{"string"}, Returns: []string{"error"}}}
+	assert.False(t, mustMatch(t, decl, r2))
+}
+
+func TestFuncDeclMatchesFilters_InvalidTypeReturnsError(t *testing.T) {
+	decl := makeFuncDecl(
+		[]*dst.Field{field(ident("string"))},
+		[]*dst.Field{field(ident("error"))},
+	)
+
+	_, err := funcDeclMatchesFilters(decl, &rule.InstFuncRule{Result: "[]invalid"}, nil)
+	require.Error(t, err)
+
+	_, err = funcDeclMatchesFilters(decl, &rule.InstFuncRule{LastResult: "[]invalid"}, nil)
+	require.Error(t, err)
+
+	_, err = funcDeclMatchesFilters(decl, &rule.InstFuncRule{Param: "[]invalid"}, nil)
+	require.Error(t, err)
+}
+
+func TestFindFuncDeclRawRule(t *testing.T) {
+	file := parseSharedFixture(t)
+
+	raw := &rule.InstRawRule{Func: "TopLevel"}
+	fn, ok, err := FindFuncDecl(file, raw)
+	require.NoError(t, err)
+	require.True(t, ok)
+	require.NotNil(t, fn)
+	assert.Equal(t, "TopLevel", fn.Name.Name)
+}
+
+func TestFindFuncDeclRawRuleMissing(t *testing.T) {
+	file := parseSharedFixture(t)
+
+	raw := &rule.InstRawRule{Func: "Missing"}
+	fn, ok, err := FindFuncDecl(file, raw)
+	require.NoError(t, err)
+	require.False(t, ok)
+	assert.Nil(t, fn)
+}
+
+func TestFindFuncDeclFilterDef(t *testing.T) {
+	file := parseSharedFixture(t)
+
+	filter := &rule.FilterDef{HasFunc: "TopLevel"}
+	fn, ok, err := FindFuncDecl(file, filter)
+	require.NoError(t, err)
+	require.True(t, ok)
+	require.NotNil(t, fn)
+	assert.Equal(t, "TopLevel", fn.Name.Name)
+}
+
+func TestFuncDeclMatchesFilters_SignatureParseError(t *testing.T) {
+	decl := makeFuncDecl(
+		[]*dst.Field{field(ident("string"))},
+		[]*dst.Field{field(ident("error"))},
+	)
+
+	_, err := funcDeclMatchesFilters(decl, &rule.InstFuncRule{
+		Signature: &rule.FuncSignature{Args: []string{"[]invalid"}},
+	}, nil)
+	require.Error(t, err)
+
+	_, err = funcDeclMatchesFilters(decl, &rule.InstFuncRule{
+		SignatureContains: &rule.FuncSignature{Args: []string{"[]invalid"}},
+	}, nil)
+	require.Error(t, err)
+
+	_, err = funcDeclMatchesFilters(decl, &rule.InstFuncRule{
+		SignatureContains: &rule.FuncSignature{Returns: []string{"[]invalid"}},
+	}, nil)
+	require.Error(t, err)
+}
+
+func TestFuncDeclMatchesFilters_LastResultNoResults(t *testing.T) {
+	decl := makeFuncDecl(nil, nil)
+
+	ok, err := funcDeclMatchesFilters(decl, &rule.InstFuncRule{LastResult: "error"}, nil)
+	require.NoError(t, err)
+	assert.False(t, ok)
+}
+
+func TestMakeAndIsUnusedIdent(t *testing.T) {
+	id := Ident("foo")
+	assert.False(t, IsUnusedIdent(id))
+
+	got := MakeUnusedIdent(id)
+	assert.Same(t, id, got)
+	assert.Equal(t, IdentIgnore, got.Name)
+	assert.True(t, IsUnusedIdent(got))
+}
+
+func TestIsStringLit(t *testing.T) {
+	lit := StringLit("hello")
+	assert.True(t, IsStringLit(lit, "hello"))
+	assert.False(t, IsStringLit(lit, "world"))
+
+	// A non-string-literal expression is never a string literal.
+	assert.False(t, IsStringLit(Ident("hello"), "hello"))
+
+	// An integer literal has the wrong Kind.
+	assert.False(t, IsStringLit(IntLit(3), "3"))
+}
+
+func TestIsInterfaceType(t *testing.T) {
+	assert.True(t, IsInterfaceType(InterfaceType()))
+	// "any" is treated as interface{}.
+	assert.True(t, IsInterfaceType(Ident("any")))
+	// Other identifiers are not interface types.
+	assert.False(t, IsInterfaceType(Ident("string")))
+}
+
+func TestIsEllipsis(t *testing.T) {
+	assert.True(t, IsEllipsis(Ellipsis(Ident("int"))))
+	assert.False(t, IsEllipsis(Ident("int")))
+}
+
+func TestAddStructField(t *testing.T) {
+	const src = `package main
+
+type S struct {
+	A int
+}
+`
+	p := NewAstParser()
+	file, err := p.ParseSource(src)
+	require.NoError(t, err)
+
+	st := FindStructType(file, "S")
+	require.NotNil(t, st)
+
+	AddStructField(st, "B", "string")
+
+	// The struct now has two fields, the new one named B of type string.
+	require.Len(t, st.Fields.List, 2)
+	newField := st.Fields.List[1]
+	require.Len(t, newField.Names, 1)
+	assert.Equal(t, "B", newField.Names[0].Name)
+	assert.Equal(t, "string", newField.Type.(*dst.Ident).Name)
+}
+
+func TestStripGenericTypes(t *testing.T) {
+	tests := []struct {
+		name string
+		expr dst.Expr
+		want string
+	}{
+		{
+			name: "value receiver",
+			expr: Ident("MyStruct"),
+			want: "MyStruct",
+		},
+		{
+			name: "pointer receiver",
+			expr: &dst.StarExpr{X: Ident("MyStruct")},
+			want: "*MyStruct",
+		},
+		{
+			name: "generic value receiver single param",
+			expr: &dst.IndexExpr{X: Ident("GenStruct"), Index: Ident("T")},
+			want: "GenStruct",
+		},
+		{
+			name: "generic pointer receiver single param",
+			expr: &dst.StarExpr{X: &dst.IndexExpr{X: Ident("GenStruct"), Index: Ident("T")}},
+			want: "*GenStruct",
+		},
+		{
+			name: "generic value receiver multiple params",
+			expr: &dst.IndexListExpr{X: Ident("GenStruct"), Indices: []dst.Expr{Ident("T"), Ident("U")}},
+			want: "GenStruct",
+		},
+		{
+			name: "generic pointer receiver multiple params",
+			expr: &dst.StarExpr{
+				X: &dst.IndexListExpr{X: Ident("GenStruct"), Indices: []dst.Expr{Ident("T"), Ident("U")}},
+			},
+			want: "*GenStruct",
+		},
+		{
+			name: "unrecognized expression yields empty",
+			expr: &dst.SelectorExpr{X: Ident("pkg"), Sel: Ident("Type")},
+			want: "",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			assert.Equal(t, tt.want, stripGenericTypes(tt.expr))
+		})
+	}
+}
